+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Update implementation of `scattering._stack_s_matrices` to use scan rather than a Python `for` loop.
 - Scalar thickness validation in scattering module.
 - Make `omega_script_k_matrix` an attribute of the `LayerSolveResult` rather than a computed property. This avoids redundant computation in cases where jit compilation is not used.
+- Provide a private `_fields_on_grid` function which does not require a full layer solve result as input.
 
 ## 1.2.0 (April 2, 2025)
 - In fields module, allow number of unit cells to be specified independently when Brillouin zone integration is used. Retain ability to infer number of unit cells from Brillouin grid axes in when number of unit cells is not specified.