+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change log
 
 ## Unreleased
+
+## 1.3.0 (April 24, 2025)
 - Add permittivity shape validation to eigensolve functions.
 - Make small changes throughout to acommodate primitive lattice vectors with batch dimensions.
 - Update vector field calculation to scan over batch, replacing prior usage of vmap. This seems to compile faster and also actually run faster.
@@ -8,6 +10,7 @@
 - Scalar thickness validation in scattering module.
 - Make `omega_script_k_matrix` an attribute of the `LayerSolveResult` rather than a computed property. This avoids redundant computation in cases where jit compilation is not used.
 - Provide a private `_fields_on_grid` function which does not require a full layer solve result as input.
+- Restructure to allow distribution of code with Meta-owned copyright under MIT License, and distribution of other code under AGPL.
 
 ## 1.2.0 (April 2, 2025)
 - In fields module, allow number of unit cells to be specified independently when Brillouin zone integration is used. Retain ability to infer number of unit cells from Brillouin grid axes in when number of unit cells is not specified.